requirements.txt
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pyyaml>=6.0
tensorboard>=2.7.0
tqdm>=4.62.0
pytest>=6.2.5
hydra-core>=1.1.0
omegaconf>=2.1.0
wandb>=0.15.0 